refactor: 서비스 패키지명 오탈자 수정
