+++ src/main/java/roomescape/controller/ReservationTimeApiController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.ReservationTimeRequest;
 import roomescape.dto.ReservationTimeResponse;
-import roomescape.servcie.ReservationTimeService;
+import roomescape.service.ReservationTimeService;
 
 @RestController
 @RequestMapping("/times")
